server/requirements.txt
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.6
cryptography==41.0.5
firebase-admin==6.2.0
Flask==2.3.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
google-api-core==2.13.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
hmmlearn==0.2.5
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
msgpack==1.0.7
numpy==1.25.2
opencv-python==4.8.0.76
pandas==2.1.0
pandus==2.1.0
proto-plus==1.22.3
protobuf==4.25.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-vlc==3.0.18122
pytz==2023.3
requests==2.31.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==2.3.6
zipp==3.16.2
